Extend the solution to all TestFinders that use SemanticDB
